backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
apscheduler==3.10.4
supabase==2.0.3
openai==1.3.0  # For DeepSeek API
tenacity==8.2.3
python-multipart==0.0.6
pydantic==2.4.2 